Form Validation 2
Added Taglib from spring framework.org/tags/form

changed the necessary form stuff to sf:form, sf:input, sf:textarea

Added Modelview to create offer method to pass a new offer.

in the sf:form added commandName=“Offer” to pass to the new createOffer
Method to create a new object

now it will stay on the page if the offer in not valid and keep all the
stuff you entered.